fix: FeeService returns int (Rupiah has no decimals), paginate casts to int
Minor fixes found during comprehensive testing:
- FeeService::calculate() now returns int instead of float
- paginate() total_pages cast to (int) for strict comparison
- Both fixes ensure === comparisons work correctly throughout the app

Full test suite: 268 checks, ALL PASSED

+++ app/Helpers.php
@@ -447,7 +447,7 @@ function dd(mixed ...$vars): never
 function paginate(array $items, int $page = 1, int $perPage = 20): array
 {
     $total = count($items);
-    $totalPages = max(1, ceil($total / $perPage));
+    $totalPages = (int)max(1, ceil($total / $perPage));
     $page = max(1, min($page, $totalPages));
     $offset = ($page - 1) * $perPage;
